condomate/backend/requirements.txt
fastapi[all]>=0.104.1
uvicorn[standard]>=0.24.0
pydantic[email]>=2.5.2
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
motor>=3.3.2
pymongo>=4.6.1
beanie>=1.23.6
aiohttp>=3.9.1

# Testing dependencies
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0
httpx>=0.25.2
tqdm>=4.66.1
colorama>=0.4.6
